+++ app_requirements.txt
@@ -1,4 +1,5 @@
-fastapi[all]==0.109.1
-uvicorn==0.22.0
-sentry-sdk[fastapi]==1.28.1
-python-decouple==3.8
+fastapi[all]
+uvicorn
+sentry-sdk[fastapi]
+python-decouple
+pydantic
