+++ deepparse/app/__init__.py
@@ -1,1 +1,3 @@
 from .app import *
+from .tools import *
+from .address import *
