+++ deepparse/app/address.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class Address(BaseModel):
+    raw: str
